Units for Total Duration
I'm not sure that a unit is needed here ("m" for "minutes"). It confused me a bit when I made a short screen recording and it said "Total Duration 0:00:12.345 m" and I thought "there is no way that recording is 12.345 minutes long".
Actually, because the first field is the hours (even though hard-coded to zero) there is some argument that the units are "hr". And the last part is seconds, so the unit could be "sec".
So I think best to not mention any unit of measure.