Improve wording of OverrideDelay IncreaseDecreaseDelay text
The use of the decimal point here (like 10.000ms) is confusing, because in European style this means 10 thousand but in US/UK style it means exactly 10 milliseconds and "000" microseconds. I guess there might be some way to localize that so that the digits separator from the locale is automagically used?
This is a suggested way to reword it, and write out the milliseconds numbers without any "." or "," separator.