835 - Settings were not being saved correctly
The parser was trying to hard to read each property and attributes.
So I added some validations, and set to save a log if something
went wrong, before parsing.

Upon save, the app was crashing if one property was set to null inside
the resource dictionary. I simple added a validation to only insert
non null entries.

Also, I increased the version number.