Parallelize capture saving to disk for ImageCapture
This changelist adds additional capture consumer tasks, which are hinted to run in parallel. These tasks speed up the encoding and saving process, which was previously a bottleneck causing large memory usage and long waiting times at the end of recording.
The degree of parallelization is set conservatively to avoid overwhelming the CPU on weaker systems.
This change is implemented for ImageCapture and DirectImageCapture, other capture methods are unchanged.